A lot of changes related to Buying and Songs Presentation

+++ frontend/src/storage/contractdef.js
@@ -106,7 +106,7 @@ smartContract =[
       },
       {
         "name": "_decimals",
-        "type": "uint256"
+        "type": "uint8"
       },
       {
         "name": "_id",
